requirements.txt
bs4==0.0.2
python-dotenv==1.1.0
requests==2.32.3
Scrapy==2.12.0
apache-airflow-providers-google
python-dotenv
